# src/main/resources/META-INF/openapi.yaml
openapi: 3.1.0
info:
  title: Vert.x Web API
  version: "1.0.0"
  description: API documentation for the Vert.x Web learning project.
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/users/{id}:
    get:
      tags: [Users]
      summary: Get user by id
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Users]
      summary: Update user
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdateRequest"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                type: object
    delete:
      tags: [Users]
      summary: Delete user
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User deleted
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    User:
      type: object
      properties:
        id:
          oneOf:
            - type: integer
            - type: string
        name:
          type: string
        email:
          type: string
        active:
          type: boolean
...
    UserUpdateRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        active:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        status:
          type: integer
        timestamp:
          oneOf:
            - type: integer
            - type: string
security:
  - bearerAuth: []